requirements.txt
curl_cffi
aiohttp
aiohttp_socks
auth
Flask[async]
g4f
Werkzeug>=3.0.3
aiohttp_socks
nodriver
pysqlite3
python-multipart
uvicorn
fastapi
python-telegram-bot==21.6